requirements.txt
Django==1.8
Markdown==2.6.2
django-extensions==1.5.2
django-filter==0.9.2
djangorestframework==3.1.1
six==1.9.0
